src/requirements.txt
bs4
icalendar
datetime
beautifulsoup4
requests
google-cloud-storage
